Return None from require if module does not exist
This will change the behaviour of fr.hmil.roshttp.node.Helpers.require
to return None if the global require method exists but the required
module does not exist.

In a browser environment packaged by scalajs-bundle the require method
may be defined in order to allow interop with common js modules, and it
returns undefined if the desired module does not exist. The result is
that RosHTTP never uses the Browser driver but fails with
scala.scalajs.js.JavaScriptException: TypeError: Cannot read property
'request' of undefined.